Refactor import paths for AppNavigator to use the correct navigation package
