+++ lib/src/shared/extensions/context_extension.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:hubx/src/app/bloc/app_bloc.dart';
 import 'package:hubx/src/app/navigation/app_navigator_impl.dart';
-import 'package:hubx/src/domain/domain/app_navigator.dart';
+import 'package:hubx/src/domain/navigation/app_navigator.dart';
 
 extension MediaQueryExtension on BuildContext {
   double get width => MediaQuery.sizeOf(this).width;
